chore: API, query 네이밍 dictionary, pet 에서 dictionaryPlant, petPlant로 변경
