+++ frontend/src/hooks/queries/petPlant/useEditPetPlant.ts
@@ -2,7 +2,7 @@ import type { EditPetPlantRequest, PetPlantDetails } from 'types/petPlant';
 import { useMutation } from '@tanstack/react-query';
 import { generatePath, useNavigate } from 'react-router-dom';
 import useAddToast from 'hooks/useAddToast';
-import PetAPI from 'apis/pet';
+import PetAPI from 'apis/petPlant';
 import noRetryIfUnauthorized from 'utils/noRetryIfUnauthorized';
 import throwOnInvalidStatus from 'utils/throwOnInvalidStatus';
 import { URL_PATH } from 'constants/index';
