+++ frontend/src/hooks/queries/petPlant/usePetPlantCardList.ts
@@ -1,29 +1,30 @@
-import type { PetPlantDetails } from 'types/petPlant';
+import type { DataResponse } from 'types/DataResponse';
+import type { PetPlantItem } from 'types/petPlant';
 import { useQuery } from '@tanstack/react-query';
-import PetAPI from 'apis/pet';
+import PetAPI, { PET } from 'apis/petPlant';
 import noRetryIfUnauthorized from 'utils/noRetryIfUnauthorized';
 import throwOnInvalidStatus from 'utils/throwOnInvalidStatus';
 import useCheckSessionId from '../auth/useCheckSessionId';
 
-const usePetPlantDetails = (petPlantId: PetPlantDetails['id']) => {
+const usePetPlantCardList = () => {
   const { isSuccess } = useCheckSessionId();
 
-  return useQuery<PetPlantDetails>({
-    queryKey: ['petPlantDetails', petPlantId],
+  return useQuery<DataResponse<PetPlantItem[]>, Error, PetPlantItem[]>({
+    queryKey: [PET, 'list'],
     queryFn: async () => {
-      const response = await PetAPI.getDetails(petPlantId);
-
+      const response = await PetAPI.getList();
       throwOnInvalidStatus(response);
 
       const data = await response.json();
       return data;
     },
 
-    refetchOnWindowFocus: false,
+    select: ({ data }) => data,
+
     suspense: true,
     retry: noRetryIfUnauthorized,
     enabled: isSuccess,
   });
 };
 
-export default usePetPlantDetails;
+export default usePetPlantCardList;
