+++ frontend/src/hooks/queries/petPlant/useRegisterPetPlant.ts
@@ -2,7 +2,7 @@ import { NewPetPlantRequest } from 'types/petPlant';
 import { useMutation } from '@tanstack/react-query';
 import { useNavigate } from 'react-router-dom';
 import useAddToast from 'hooks/useAddToast';
-import PetAPI from 'apis/pet';
+import PetAPI from 'apis/petPlant';
 import noRetryIfUnauthorized from 'utils/noRetryIfUnauthorized';
 import throwOnInvalidStatus from 'utils/throwOnInvalidStatus';
 import { URL_PATH } from 'constants/index';
