+++ frontend/src/pages/PetPlantCardList/index.tsx
@@ -2,7 +2,7 @@ import { Link, generatePath } from 'react-router-dom';
 import Navbar from 'components/@common/Navbar';
 import PetCard from 'components/petPlant/PetPlantCard';
 import { CardList, RegisterButton, Title, Wrapper } from './PetPlantCardList.style';
-import usePetPlantCardList from 'hooks/queries/pet/usePetPlantCardList';
+import usePetPlantCardList from 'hooks/queries/petPlant/usePetPlantCardList';
 import { URL_PATH } from 'constants/index';
 
 const PetPlantCardList = () => {
