+++ frontend/src/types/dictionaryPlant.ts
@@ -15,7 +15,7 @@ export interface DictionaryPlant {
   waterCycle: Record<Season, string>;
 }
 
-export type DictNameSearchResult = Pick<DictionaryPlant, 'id' | 'name' | 'image'>;
+export type DictionaryPlantNameSearchResult = Pick<DictionaryPlant, 'id' | 'name' | 'image'>;
 
 export type Season = 'spring' | 'summer' | 'autumn' | 'winter';
 export type SeasonKor = '봄' | '여름' | '가을' | '겨울';
